Sketching intensifies.
File pulled out into interface; Dirs have some sketched methods now;
added an Openmode enum (that's... gonna need work); add a DirItr type
(that might grow or have a sibling interface that lets the user have
more control of bulk reads, for efficiently/alloccontrol; we'll see);
and so on...

A fundamental question occurs in a couple of places:
Are we going to have this API admit to and be shaped by what operations
need individual syscalls (and defacto, define that as "on linux")?

It's good to design APIs to help users DTRT regarding performance.
It's also tricky to define error handling paths if we try to
expose something as if it's an atomic op when in fact it's not.

On the other hand... some of the lines are terribly arbitrary;
and there's only certain safe ways and orders to compose them
(consider: chown, setsuid, and setmtime -- only some orderings of
these operations have the effect intended!), and it'd be great to
be able to save users from being exposed to that sideshow.

Tricky business, this whole filesystems thing.