outline some move operations, and comments.
I've come to the realization that a lot of what the go standard lib
exposes is kinda just barely scratching the surface of what one can
actually do when one goes to the actual kernel ABI.

The question of what to do with this... remains.  I do still want this
package to be reasonably platform agnostic.